Advanced changes to mmd
Allows some of the features to be used in other avatars
