+++ hifi_tools/utils/mmd_hifi_dict.csv
@@ -148,6 +148,7 @@
 "ヴィーナス", "Venus", "vinasu",
 "恐ろしい子", "HorrorChild", "osoroshi ko",
 "蝶ネクタイ", "BowTie", "chounekutai",
+"ホイッスル", "Whistle", "hoissuru",
 "ガーター", "Garter", "gata",
 "ジップ", "Zip", "zippu",
 "Root", "Root",
@@ -192,6 +193,7 @@
 "グループ", "Group", "guruuppu",
 "サンダル", "Sandals", "sandaru",
 "シエスタ", "Siesta", "shiesuta",
+"シュシュ", "Chouchu", "shushu",
 "ショーツ", "Shorts", "sho-tsu",
 "ジーンズ", "Jean", "ji-nzu",
 "スカト横", "SkirtSide",
@@ -245,11 +247,14 @@
 "服色調整", "CostumeColor", "fuku shoku chōsei",
 "縦れ防止", "PreventSagging", "taterebotome",
 "胸押さえ", "Bust_guard",
+"ホルダー", "Holder", "horuda",
 "ケモミミ", "Ears", "kemomimi",
 "透ジャ", "Translucent", "torushi?",
 "鎖移動用", "ChainMovement",
 "髪色調整", "HairColor", "kami shoku chōsei",
 "オドロキ", "Surprise", "odoroki",
+"ウインク", "Wink", "uinku",
+"グローブ", "Glove", "gurobu",
 "足先EX", "Toe"
 "ポケット", "Pocket", "poketto",
 "サビオ", "Bandage", "sahi o",
@@ -263,6 +268,7 @@
 "うるる", "Cry", "uru uru",
 "うわぎ", "Coat", "uwagi",
 "おさげ", "Pigtail", "osage",
+"かかと", "Heel", "kakato",
 "がおー", "Roar",
 "がーん", "Gloom", "gaan",
 "きいろ", "Yellow", "kiiro",
@@ -273,6 +279,7 @@
 "せびろ", "Suit", "sebiro",
 "だくみ", "Mechanism", "da kumi",
 "てへっ", "TehHeh", "tehheh",
+"レンズ", "Lens", "renzu",
 "とほほ", "Tears",
 "なごみ", "Calm", "nagomi",
 "なぬっ", "Coloring",
@@ -290,6 +297,7 @@
 "ぺろっ", "Lick",
 "ぼうし", "Hat", "boushi",
 "まつげ", "Eyelash", "matsuge",
+"まつ毛", "Eyelash", "matsuge",
 "みずぎ", "Swimsuit", "mizugi",
 "みどり", "Green", "midori",
 "めがね", "Glasses", "megane",
@@ -379,6 +387,7 @@
 "真面目", "Serious",
 "神奈子", "Kanako",
 "移動用", "Move", "ido?",
+"ラバー", "Rubber", "raba",
 "紐ぱん", "String",
 "細める", "Tighten", "hosomeru",
 "絆創膏", "Bandage", "",
@@ -389,14 +398,18 @@
 "衣料品", "Clothing", "iryouhin",
 "西行寺", "Saigyouji",
 "見開く", "Spread", "mihiraku",
+"メタル", "Metal", "metaru",
 "赤寄り", "Red", "aka yori",
 "鎖移動", "ChainMovement", "kusari ido",
 "青ざめ", "Pale", "aozame",
 "青寄り", "Blue", "ao yori",
+"ホルダ", "Holder", "horuda",
+"つなぎ", "Tether", "tsunagi",
 "革ぱん", "Leather",
 "キュッ", "Cute", "Kyu-tsu",
 "穏やか", "Gentle", "odayaka",
 "髪留め", "RubberBand",
+"ライン", "Line", "rain",
 "ジト目", "Iris", "jito-me",
 "その他", "Other", "",
 "体(下)", "Lower Body", "",
@@ -407,6 +420,7 @@
 "ケット", "Jacket", "ketto",
 "レオタ", "Leotard", "reota",
 "レグ", "UpLeg", "regu",
+"ヒィー", "Hey", "hyi",
 "ぺろっ", "Licking", "",
 "ω□", "Omegabox",
 "瞳小", "Pupil Small", "",
@@ -556,11 +570,14 @@
 "着用", "Wear", "chakuyo",
 "瞳孔", "Pupil", "dookoo",
 "瞳小", "Pupil",
+"不安", "Anxiety", "fuan",
 "移動", "Move", "ido",
 "笑い", "Smile", "warai",
 "笑う", "Laugh",
 "細い", "Thinner",
 "細眉", "Thin Eyebrows",
+"ギザ", "Giza", "giza",
+"わぁ", "Wow", "wa-a",
 "紳士", "Gentleman", "shinshi",
 "紺色", "Navy_blue", "koniro",
 "経巻", "Scroll", "kyōkan",
@@ -615,6 +632,7 @@
 "靴下", "Socks", "kutsushita",
 "風見", "Kazami",
 "飾り", "Accessory", "kazari",
+"ひも", "String", "himo",
 "鳶色", "Reddish_brown", "tobiiro",
 "黄色", "Yellow", "kiiro",
 "黒い", "Black", "kuroi",
@@ -839,6 +857,10 @@
 "木", "Wood", "ki",
 "規", "Control", "tadashi",
 "板", "Board", "ita",
+"皮", "Skin", "pi",
+"リ", "ri", "ri",
+"★", "", "",
+"コ", "co", "co",
 "！", "!",
 "＋", "+",
 "－", "-",
